move stack describe to its own file
